Experience fix

+++ backend/src/main/java/com/example/backend_v2/repositories/ExperienceRepository.java
@@ -3,7 +3,9 @@ package com.example.backend_v2.repositories;
 import com.example.backend_v2.dao.ExperienceDao;
 import com.example.backend_v2.models.Company;
 import com.example.backend_v2.models.Experience;
+import com.example.backend_v2.models.ExperienceAndCompany;
 import com.example.backend_v2.repositories.rowMappers.CompanyRowMapper;
+import com.example.backend_v2.repositories.rowMappers.ExperienceAndCompanyRowMapper;
 import com.example.backend_v2.repositories.rowMappers.ExperienceRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -21,21 +23,18 @@ public class ExperienceRepository implements ExperienceDao {
     }
 
     @Override
-    public List<Experience> getExperiences(int user_id) {
+    public List<ExperienceAndCompany> getExperiences(int user_id) {
         var sql = """
-                SELECT e.exp_ID AS exp_ID, e.name AS exp_name, e.type AS type, e.title 
-                AS title, e.start_date AS start_date, e.end_date AS end_date, c.name AS
-                company_name
+                SELECT *
                 FROM (Experience NATURAL JOIN Exp_company) AS e JOIN Company
                 AS c ON e.company_ID = c.company_ID
-                WHERE user_ID = ? AND e.type != 'EDUCATION'
+                WHERE e.user_ID = ? AND e.type != 'EDUCATION'
                 """;
-        return jdbcTemplate.query(sql, new ExperienceRowMapper(), user_id);
+        return jdbcTemplate.query(sql, new ExperienceAndCompanyRowMapper(), user_id);
     }
 
     @Override
     public int insertExperience(Experience experience, String companyName) {
-        System.out.println(experience.title() + "\n\n\n\n\n\n");
         var sql = """
                 INSERT INTO Experience (exp_ID, user_ID, title, description, type, start_date, end_date)
                 VALUES (?, ?, ?, ?, 'WORK', ?, ?);
@@ -44,7 +43,7 @@ public class ExperienceRepository implements ExperienceDao {
 
 
         sql = """
-                SELECT company_ID FROM Company WHERE name = ?
+                SELECT * FROM Company WHERE name = ?
                 """;
         Optional<Company> temp = jdbcTemplate.query(sql, new CompanyRowMapper(), companyName).stream().findFirst();
 
@@ -65,38 +64,51 @@ public class ExperienceRepository implements ExperienceDao {
     @Override
     public int editExperience(Experience experience, String companyName) {
         var sql = """
+                SELECT * FROM Company WHERE name = ?
+                """;
+        Optional<Company> temp = jdbcTemplate.query(sql, new CompanyRowMapper(), companyName).stream().findFirst();
+
+        sql = """
                 UPDATE Experience
                 SET title = ?, description = ?, start_date = ?, end_date = ?
                 WHERE exp_ID = ? AND user_ID = ? AND type != 'EDUCATION';
                 UPDATE Exp_company
-                SET company_ID = SELECT company_ID FROM Company WHERE name = ?
+                SET company_ID = ?
                 WHERE exp_ID = ?;
                 """;
-        return jdbcTemplate.update(sql, experience.title(), experience.description(), experience.start_date(), experience.end_date(), experience.exp_ID(), experience.user_ID(), companyName, experience.exp_ID());
+        return jdbcTemplate.update(sql, experience.title(), experience.description(), experience.start_date(), experience.end_date(), experience.exp_ID(), experience.user_ID(), temp.get().company_ID(), experience.exp_ID());
     }
 
     @Override
     public int deleteExperience(int exp_id, int user_id) {
         var sql = """
-                DELETE FROM Experience
-                WHERE exp_ID = ? AND user_ID = ? AND type != 'EDUCATION';
+                SELECT * FROM Experience
+                WHERE exp_ID = ? AND user_ID = ?;
+                """;
+        Optional<Experience> experience = jdbcTemplate.query(sql, new ExperienceRowMapper(), exp_id, user_id).stream().findFirst();
+        if(experience.isPresent() && experience.get().type().equals("EDUCATION")) {
+            return -1;
+        }
+        sql = """
                 DELETE FROM Exp_company
                 WHERE exp_ID = ?;
+                DELETE FROM experiences
+                WHERE exp_ID = ?;
+                DELETE FROM Experience
+                WHERE exp_ID = ? AND user_ID = ? AND type != 'EDUCATION';
                 """;
-        return jdbcTemplate.update(sql, exp_id, user_id, exp_id);
+        return jdbcTemplate.update(sql, exp_id, exp_id, exp_id, user_id);
     }
 
     @Override
-    public List<Experience> getEducations(int user_id) {
+    public List<ExperienceAndCompany> getEducations(int user_id) {
         var sql = """
-                SELECT e.exp_ID AS exp_ID, e.name AS exp_name, e.title AS title,
-                e.start_date AS start_date, e.end_date AS end_date, c.name AS
-                company_name
+                SELECT *
                 FROM (Experience NATURAL JOIN Exp_company) AS e JOIN Company
                 AS c ON e.company_ID = c.company_ID
-                WHERE user_ID = ? AND e.type = 'EDUCATION'
+                WHERE e.user_ID = ? AND e.type = 'EDUCATION'
                 """;
-        return jdbcTemplate.query(sql, new ExperienceRowMapper(), user_id);
+        return jdbcTemplate.query(sql, new ExperienceAndCompanyRowMapper(), user_id);
     }
 
     @Override
@@ -104,33 +116,64 @@ public class ExperienceRepository implements ExperienceDao {
         var sql = """
                 INSERT INTO Experience (exp_ID, user_ID, title, description, type, start_date, end_date)
                 VALUES (?, ?, ?, ?, 'EDUCATION', ?, ?);
-                INSERT INTO Exp_company (exp_ID, company_ID)
-                VALUES (?, SELECT company_ID FROM Company WHERE name = ?);
                 """;
-        return jdbcTemplate.update(sql, experience.exp_ID(), experience.user_ID(), experience.title(), experience.description(), experience.start_date(), experience.end_date(), experience.exp_ID(), institution);
+        jdbcTemplate.update(sql, experience.exp_ID(), experience.user_ID(), experience.title(), experience.description(), experience.start_date(), experience.end_date());
+
+
+        sql = """
+                SELECT * FROM Company WHERE name = ?
+                """;
+        Optional<Company> temp = jdbcTemplate.query(sql, new CompanyRowMapper(), institution).stream().findFirst();
+
+        sql = """
+                INSERT INTO Exp_company (exp_ID, user_ID, company_ID)
+                VALUES (?, ?, ?);
+                """;
+        jdbcTemplate.update(sql, experience.exp_ID(), experience.user_ID(), temp.get().company_ID());
+
+        sql = """
+                INSERT INTO experiences (exp_ID, user_ID)
+                VALUES (?, ?);
+                """;
+        return jdbcTemplate.update(sql, experience.exp_ID(), experience.user_ID());
     }
 
     @Override
     public int editEducation(Experience experience, String institution) {
         var sql = """
+                SELECT * FROM Company WHERE name = ?
+                """;
+        Optional<Company> temp = jdbcTemplate.query(sql, new CompanyRowMapper(), institution).stream().findFirst();
+
+        sql = """
                 UPDATE Experience
                 SET title = ?, description = ?, start_date = ?, end_date = ?
                 WHERE exp_ID = ? AND user_ID = ? AND type = 'EDUCATION';
                 UPDATE Exp_company
-                SET company_ID = SELECT company_ID FROM Company WHERE name = ?
+                SET company_ID = ?
                 WHERE exp_ID = ?;
                 """;
-        return jdbcTemplate.update(sql, experience.title(), experience.description(), experience.start_date(), experience.end_date(), experience.exp_ID(), experience.user_ID(), institution, experience.exp_ID());
+        return jdbcTemplate.update(sql, experience.title(), experience.description(), experience.start_date(), experience.end_date(), experience.exp_ID(), experience.user_ID(), temp.get().company_ID(), experience.exp_ID());
     }
 
     @Override
     public int deleteEducation(int exp_id, int user_id) {
         var sql = """
-                DELETE FROM Experience
-                WHERE exp_ID = ? AND user_ID = ? AND type = 'EDUCATION';
+                SELECT * FROM Experience
+                WHERE exp_ID = ? AND user_ID = ?;
+                """;
+        Optional<Experience> experience = jdbcTemplate.query(sql, new ExperienceRowMapper(), exp_id, user_id).stream().findFirst();
+        if(experience.isPresent() && !experience.get().type().equals("EDUCATION")) {
+            return -1;
+        }
+        sql = """
                 DELETE FROM Exp_company
                 WHERE exp_ID = ?;
+                DELETE FROM experiences
+                WHERE exp_ID = ?;
+                DELETE FROM Experience
+                WHERE exp_ID = ? AND user_ID = ? AND type = 'EDUCATION';
                 """;
-        return jdbcTemplate.update(sql, exp_id, user_id, exp_id);
+        return jdbcTemplate.update(sql, exp_id, exp_id, exp_id, user_id);
     }
 }
